+++ plugin-api/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @matatbread/matbot-plugin-api
 
+## 0.4.10
+
 ## 0.4.9
 
 ## 0.4.8